Cleanup (#77)
* Cleanup

* Cleanup

* Cleanup
