+++ app/api/urls.py
@@ -1,4 +1,4 @@
-from django.urls import path
+from django.urls import path, re_path
 
 from api import views
 
@@ -6,6 +6,8 @@ app_name = 'api'
 
 urlpatterns = [
     path('', views.api_view, name='status'),
+    re_path(r'^upload/?$', views.upload_view, name='upload'),
+    re_path(r'^shorten/?$', views.shorten_view, name='shorten'),
     path('recent/', views.recent_view, name='recent'),
     path('remote/', views.remote_view, name='remote'),
     path('stats/', views.stats_view, name='stats'),
