+++ app/home/urls.py
@@ -1,5 +1,6 @@
 from django.urls import path, re_path
 
+from api.views import shorten_view, upload_view
 from home import views
 
 app_name = 'home'
@@ -12,10 +13,8 @@ urlpatterns = [
     path('shorts/', views.shorts_view, name='shorts'),
     path('settings/', views.settings_view, name='settings'),
     path('stats/', views.stats_view, name='stats'),
-    re_path(r'^upload/?$', views.upload_view, name='upload'),
-    re_path(r'^api/upload/?$', views.upload_view, name='api-upload'),
-    re_path(r'^shorten/?$', views.shorten_view, name='shorten'),
-    re_path(r'^api/shorten/?$', views.shorten_view, name='api-shorten'),
+    re_path(r'^upload/?$', upload_view, name='upload'),
+    re_path(r'^shorten/?$', shorten_view, name='shorten'),
     path('s/<str:short>', views.shorten_short_view, name='short'),
     path('ajax/update/stats/', views.update_stats_ajax, name='update-stats'),
     path('ajax/delete/file/<int:pk>/', views.delete_file_ajax, name='delete-file'),
