+++ app/home/util/s3.py
@@ -1,4 +1,3 @@
-
 import os
 from storages.backends.s3boto3 import S3Boto3Storage
