+++ app/oauth/admin.py
@@ -1,7 +1,7 @@
 from django.contrib import admin
 from django.contrib.auth.admin import UserAdmin
 
-from .models import CustomUser
+from oauth.models import CustomUser
 
 
 @admin.register(CustomUser)
