+++ app/oauth/forms.py
@@ -1,7 +1,7 @@
 from django import forms
 from django.contrib.auth.forms import UserCreationForm, UserChangeForm
 
-from .models import CustomUser
+from oauth.models import CustomUser
 
 
 class CustomUserCreationForm(UserCreationForm):
