+++ app/oauth/urls.py
@@ -1,6 +1,6 @@
 from django.urls import path
 
-from . import views
+from oauth import views
 
 
 app_name = 'oauth'
