+++ app/requirements.txt
@@ -23,5 +23,5 @@ pytimeparse2
 sentry-sdk
 uvicorn[standard]
 validators
-# watchdog
+watchdog
 Werkzeug
